changed folder and file names + removed unneeded functions in harr cascade + fixed reading video for checking if not video is found
