+++ HaarCascadeLipsDetector/FramesManipulations.py
@@ -1,0 +1,26 @@
+import cv2
+import numpy as np
+# ----------------------------------------------------------------------------
+# function used to resize image
+# input: image, dim = (x,y)
+# output: image
+def resizeImage(img, dim=(650, 650)):
+    resized = cv2.resize(img, dim, interpolation=cv2.INTER_AREA)
+    return resized
+# ----------------------------------------------------------------------------
+# function used to extract frames from video
+# input: video
+# output: frames
+def getVideoFrames(videoPath):
+    """Function to return a video's frames in a list
+    :type videoPath: String
+    """
+    vidcap = cv2.VideoCapture(videoPath)
+    if vidcap.isOpened() == False:
+        return [], False
+    success, image = vidcap.read()
+    allFrames = []
+    while success:
+        allFrames.append(image)
+        success, image = vidcap.read()
+    return allFrames, True
